+++ package.json
@@ -5,10 +5,10 @@
   "license": "MIT",
   "main": "lib/levelup.js",
   "scripts": {
-    "test": "standard && hallmark && (nyc -s node test | faucet) && nyc report",
+    "test": "standard && hallmark && (nyc -s node test/self.js | faucet) && nyc report",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "test-browsers": "airtap --loopback airtap.local test/index.js",
-    "test-browser-local": "airtap --local test/index.js",
+    "test-browsers": "airtap --loopback airtap.local test/self.js",
+    "test-browser-local": "airtap --local test/self.js",
     "hallmark": "hallmark --fix",
     "dependency-check": "dependency-check . test/*.js",
     "prepublishOnly": "npm run dependency-check"
